chore(deps): bump every package to latest; hold typescript 5.9, babel 7, @types/bun 1.3.14 and claude-code 2.1.45
Blanket bun update --latest in the root and src/utils workspaces, then the code moved to the new APIs:
markdown-it 15 ships its own types (Token/Env/MarkdownIt named exports, no lib/token.mjs),
@mdit/plugin-alert 2 renamed openRender/closeRender/titleRender to *Renderer, lucide-react 1
dropped brand icons (Youtube -> Tv), kysely 0.29 wants sql<SqlBool> in a join predicate,
tar-stream 3.2.1 types its streams via streamx (bumped to 2.28 for the d.ts), and
@opentui/core 0.5 packs RGBA into a Uint16Array.

Held back on purpose: typescript 7 is the Go port without the compiler API mcp-tsc uses;
@babel/core 8 breaks babel-preset-solid (peer ^7); @types/bun 1.4.0 hides every
process.on(signal) overload behind its memoryPressure augmentation (25 type errors, same under
tsc); @anthropic-ai/claude-code >= 2.1.185 is a native-binary wrapper without the vendored
ripgrep that src/utils/ripgrep.ts falls back to.

+++ src/doctor/ui/tui/views/__tests__/drawer-content-packing.test.ts
@@ -43,7 +43,7 @@ describe("drawer content packing", () => {
 
         const rgba = toRgba("#7aa2f7");
         expect(rgba).toBeInstanceOf(RGBA);
-        expect(rgba?.buffer).toBeInstanceOf(Float32Array);
+        expect(rgba?.buffer).toBeInstanceOf(Uint16Array);
         expect(rgba?.buffer.length).toBe(4);
     });
 
@@ -69,12 +69,12 @@ describe("drawer content packing", () => {
 
                     if (chunk.fg !== undefined) {
                         expect(chunk.fg).toBeInstanceOf(RGBA);
-                        expect((chunk.fg as RGBA).buffer).toBeInstanceOf(Float32Array);
+                        expect((chunk.fg as RGBA).buffer).toBeInstanceOf(Uint16Array);
                     }
 
                     if (chunk.bg !== undefined) {
                         expect(chunk.bg).toBeInstanceOf(RGBA);
-                        expect((chunk.bg as RGBA).buffer).toBeInstanceOf(Float32Array);
+                        expect((chunk.bg as RGBA).buffer).toBeInstanceOf(Uint16Array);
                     }
                 }
             }
